requirements.txt
dkimpy~=1.0.5